Move all public types from NetMQ.zmq namespace into NetMQ.
Closes #299.
