+++ src/NetMQ.Tests/ReceivingSocketExtensionsTests.cs
@@ -3,7 +3,6 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using JetBrains.Annotations;
-using NetMQ.zmq;
 using NUnit.Framework;
 
 namespace NetMQ.Tests
