+++ src/NetMQ.Tests/SocketTests.cs
@@ -5,7 +5,6 @@ using System.Net.Sockets;
 using System.Text;
 using System.Threading;
 using System.Threading.Tasks;
-using NetMQ.zmq;
 using NUnit.Framework;
 
 // ReSharper disable ExceptionNotDocumented
