+++ src/NetMQ/Endianness.cs
@@ -1,0 +1,12 @@
+﻿namespace NetMQ
+{
+    /// <summary>
+    /// This enum-type specifies either big-endian (Big) or little-endian (Little),
+    /// which indicate whether the most-significant bits are placed first or last in memory.
+    /// </summary>
+    public enum Endianness
+    {
+        Big,
+        Little
+    }
+}
