+++ src/NetMQ/IReceivingSocket.cs
@@ -1,5 +1,4 @@
 ﻿using System;
-using NetMQ.zmq;
 
 namespace NetMQ
 {
