+++ src/NetMQ/NetMQActor.cs
@@ -2,7 +2,6 @@
 using System.Threading;
 using JetBrains.Annotations;
 using NetMQ.Sockets;
-using NetMQ.zmq;
 
 namespace NetMQ
 {
