+++ src/NetMQ/NetMQSocketEventArgs.cs
@@ -1,6 +1,5 @@
 ﻿using System;
 using JetBrains.Annotations;
-using NetMQ.zmq;
 
 namespace NetMQ
 {
