+++ src/NetMQ/Proxy.cs
@@ -1,7 +1,6 @@
 ﻿using System;
 using System.Threading;
 using JetBrains.Annotations;
-using NetMQ.zmq;
 
 namespace NetMQ
 {
