+++ src/NetMQ/ReceivingSocketExtensions.cs
@@ -5,7 +5,6 @@ using System.Diagnostics.CodeAnalysis;
 using System.Linq;
 using System.Text;
 using JetBrains.Annotations;
-using NetMQ.zmq;
 
 namespace NetMQ
 {
