+++ src/NetMQ/ZmqNamespacePlaceholder.cs
@@ -1,0 +1,14 @@
+﻿using System;
+
+// ReSharper disable CheckNamespace
+// ReSharper disable UnusedMember.Global
+
+namespace NetMQ.zmq
+{
+    /// <summary>
+    /// All public types have been moved out of the <see cref="NetMQ.zmq"/> namespace. This class remains
+    /// to prevent compile errors when users still have <c>using NetMQ.zmq;</c> in their code.
+    /// </summary>
+    [Obsolete("Don't use this type!")]
+    public static class ThisNamespaceIsNotUsed { }
+}
