+++ src/NetMQ/zmq/Enums.cs
@@ -12,74 +12,6 @@ namespace NetMQ.zmq
     }
 
     /// <summary>
-    /// This enum-type is used to specify the basic type of message-queue socket
-    /// based upon the intended pattern, such as Pub,Sub, Req,Rep, Dealer,Router, Pull,Push, Xpub,Xsub.
-    /// </summary>
-    public enum ZmqSocketType
-    {
-        /// <summary>
-        /// No socket-type is specified
-        /// </summary>
-        None = -1,
-
-        /// <summary>
-        /// This denotes a Pair socket (usually paired with another Pair socket).
-        /// </summary>
-        Pair = 0,
-
-        /// <summary>
-        /// This denotes a Publisher socket (usually paired with a Subscriber socket).
-        /// </summary>
-        Pub = 1,
-
-        /// <summary>
-        /// This denotes a Subscriber socket (usually paired with a Publisher socket).
-        /// </summary>
-        Sub = 2,
-
-        /// <summary>
-        /// This denotes a Request socket (usually paired with a Response socket).
-        /// </summary>
-        Req = 3,
-
-        /// <summary>
-        /// This denotes a Response socket (usually paired with a Request socket).
-        /// </summary>
-        Rep = 4,
-
-        /// <summary>
-        /// This denotes an Dealer socket (CBL)
-        /// </summary>
-        Dealer = 5,
-
-        /// <summary>
-        /// This denotes an Router socket (CBL)
-        /// </summary>
-        Router = 6,
-
-        /// <summary>
-        /// This denotes a Pull socket (usually paired with a PUsh socket).
-        /// </summary>
-        Pull = 7,
-
-        /// <summary>
-        /// This denotes a Push socket (usually paired with a Pull socket).
-        /// </summary>
-        Push = 8,
-
-        /// <summary>
-        /// This denotes an XPublisher socket (CBL)
-        /// </summary>
-        Xpub = 9,
-
-        /// <summary>
-        /// This denotes an XSubscriber socket (CBL)
-        /// </summary>
-        Xsub = 10,
-        Stream = 11
-    }
-
-    /// <summary>
     /// This enum-type serves to identity a particular socket-option.
     /// </summary>
     internal enum ZmqSocketOption
@@ -166,103 +98,5 @@ namespace NetMQ.zmq
         /// Specifies the byte-order: big-endian, vs little-endian.
         /// </summary>
         Endian = 1000
-    }
-
-    /// <summary>
-    /// This enum-type specifies either big-endian (Big) or little-endian (Little),
-    /// which indicate whether the most-significant bits are placed first or last in memory.
-    /// </summary>
-    public enum Endianness
-    {
-        Big,
-        Little
-    }
-
-    /// <summary>
-    /// This flags enum-type provides a way to specify basic Receive behaviour.
-    /// It may be None, or have the DontWait bit (indicating to wait for a message),
-    /// or the SendMore bit, set.
-    /// </summary>
-    [Flags]
-    public enum SendReceiveOptions
-    {
-        /// <summary>
-        /// Both bits cleared (neither DontWait nor SendMore are set)
-        /// </summary>
-        None = 0,
-
-        /// <summary>
-        /// Set this flag to specify NOT to block waiting for a message to arrive.
-        /// </summary>
-        DontWait = 1,
-
-        /// <summary>
-        /// Set this (the SendMore bit) to signal more messages beyond the current one, for a given unit of communication.
-        /// </summary>
-        SendMore = 2,
-
-        // Deprecated aliases
-        [Obsolete("Use DontWait instead")]
-        NoBlock = DontWait
-    }
-
-    // Socket transport events (tcp and ipc only)
-
-    /// <summary>
-    /// This enum-type specifies socket transport events (TCP and IPC only).
-    /// </summary>
-    [Flags]
-    public enum SocketEvent
-    {
-        Connected = 1,
-        ConnectDelayed = 2,
-        ConnectRetried = 4,
-
-        Listening = 8,
-        BindFailed = 16,
-
-        Accepted = 32,
-        AcceptFailed = 64,
-
-        Closed = 128,
-        CloseFailed = 256,
-        Disconnected = 512,
-
-        All = Connected | ConnectDelayed |
-              ConnectRetried | Listening |
-              BindFailed | Accepted |
-              AcceptFailed | Closed |
-              CloseFailed | Disconnected
-    }
-
-    /// <summary>
-    /// This flags enum-type is simply an indication of the direction of the poll-event,
-    /// and can be None, PollIn, PollOut, or PollError.
-    /// </summary>
-    [Flags]
-    public enum PollEvents
-    {
-        None = 0x0,
-        PollIn = 0x1,
-        PollOut = 0x2,
-        PollError = 0x4
-    }
-
-    public static class PollEventsExtensions
-    {
-        public static bool HasIn(this PollEvents pollEvents)
-        {
-            return (pollEvents & PollEvents.PollIn) == PollEvents.PollIn;
-        }
-
-        public static bool HasOut(this PollEvents pollEvents)
-        {
-            return (pollEvents & PollEvents.PollOut) == PollEvents.PollOut;
-        }
-
-        public static bool HasError(this PollEvents pollEvents)
-        {
-            return (pollEvents & PollEvents.PollError) == PollEvents.PollError;
-        }
     }
 }
